maritime-ops-phase0/requirements.txt
google-generativeai>=0.8.0
pydantic>=2.0.0
pyyaml>=6.0
python-dateutil>=2.8.0
